refactor: isolate manual task batch settlement hook
